separates out bom and artifact-tree tests
