+++ tests/bom.rs
@@ -40,25 +40,3 @@ fn bom_output_test() -> Result<(), Box<dyn std::error::Error>> {
         .stdout(predicate::str::contains("Generated GitOid: 5b2f2d4e79e6387ca9dedad500ebf70e9fb3097773252cc5b9a6d5a35a987028"));
     Ok(())
 }
-
-#[test]
-fn artifact_tree_output_file_does_not_exist() -> Result<(), Box<dyn std::error::Error>> {
-    let mut cmd = Command::cargo_bin("gitbom-cli")?;
-    cmd.arg("artifact-tree").arg("directory/does/not/exist");
-    cmd.assert()
-        .failure()
-        .stderr(predicate::str::contains("No such file or directory"));
-    Ok(())
-}
-
-#[test]
-fn artifact_tree_output_test() -> Result<(), Box<dyn std::error::Error>> {
-    let mut cmd = Command::cargo_bin("gitbom-cli")?;
-    cmd.arg("artifact-tree").arg("tests/fixtures/directory_thing");
-    cmd.assert()
-        .success()
-        .stdout(predicate::str::contains("Generated GitBom for 2 files"));
-    Ok(())
-}
-
-
